pyproject.toml
[project]
name = "url-to-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.7.8",
    "fastapi[standard]>=0.128.0",
    "openai>=2.14.0",
    "pydantic-ai>=1.40.0",
    "python-dotenv>=1.2.1",
]
